# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Lockpick game.",
        "title": "Lockpick API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/games": {
            "post": {
                "description": "Create a new game room, optionally private or public for matchmaking",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Create a new game",
                "parameters": [
                    {
                        "description": "Game configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.CreateGameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.CreateGameResponse"
                        }
                    }
                }
            }
        },
        "/games/join": {
            "post": {
                "description": "Join a game by room ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Join an existing game",
                "parameters": [
                    {
                        "description": "Join parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.JoinGameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.JoinGameResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check if the server is healthy",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health Check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "server.CreateGameRequest": {
            "type": "object",
            "properties": {
                "config": {
                    "$ref": "#/definitions/store.GameConfig"
                },
                "player_name": {
                    "type": "string"
                }
            }
        },
        "server.CreateGameResponse": {
            "type": "object",
            "properties": {
                "config": {
                    "$ref": "#/definitions/store.GameConfig"
                },
                "player_id": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "server.JoinGameRequest": {
            "type": "object",
            "properties": {
                "player_name": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                }
            }
        },
        "server.JoinGameResponse": {
            "type": "object",
            "properties": {
                "config": {
                    "$ref": "#/definitions/store.GameConfig"
                },
                "player_id": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "store.GameConfig": {
            "type": "object",
            "properties": {
                "hints_enabled": {
                    "type": "boolean"
                },
                "is_private": {
                    "type": "boolean"
                },
                "pin_length": {
                    "type": "integer"
                },
                "player_name": {
                    "type": "string"
                },
                "timer_duration": {
                    "description": "0 means no timer",
                    "type": "integer"
                }
            }
        }
    }
}